feat: Create new date converter to avoid saving not needed times
