+++ lib/core/database/app_db.dart
@@ -8,8 +8,9 @@ import 'package:monekin/core/database/services/category/category_service.dart';
 import 'package:monekin/core/database/services/currency/currency_service.dart';
 import 'package:monekin/core/database/services/user-setting/user_setting_service.dart';
 import 'package:monekin/core/database/sql/initial/seed.dart';
-import 'package:monekin/core/database/utils/database_enum.dart';
-import 'package:monekin/core/database/utils/list_converters.dart';
+import 'package:monekin/core/database/utils/converters/custom_enum_converter.dart';
+import 'package:monekin/core/database/utils/converters/date_converters.dart';
+import 'package:monekin/core/database/utils/converters/list_converters.dart';
 import 'package:monekin/core/models/account/account.dart';
 import 'package:monekin/core/models/asset/asset.dart';
 import 'package:monekin/core/models/budget/budget.dart';
